[amber] generate root metadata 1 through 5
According to the TUF 1.0 Spec, we must have continguous root metadata
in our repositories, but we were currently only storing version 6. This
generates versions 1 through 5 of the metadata that just replicates what
is in version 6. rust-tuf initializes it's trust against a specific
version, but we don't yet have a way of configuring this version, so the
safest way to establish trust is to start with version 1. These root
metadata in these files are identical to version 6, and so rust-tuf should
be able to walk through these files to get up to date.

This was generated by modifying the current version 6 root metadata to
an older version, then re-signing it with go-tuf's `tuf sign root.json`.

Fixed: 39645
Change-Id: Ifec84dbbaeac3c71914157eb5cb797cf1a6e273e