[wlan][mlme] Always call SetChannel
This code previously would not call WlanmacProxy::SetChannel if it
determined that the Device's internal state's channel is equal to the
one being set. However, this does not work in AP mode on ath10k, as
setting the channel apparently also clears some internal state that
would otherwise cause the driver to crash if the channel is not set
again.

I believe the correct behavior in this case is to just always call
SetChannel and not try to optimize for the already-set case: the driver
should handle this or, if no-opping the set is actually required, to do
this at the MLME level and ignore the Device's internal state.

Test: tested on ath10k
Change-Id: I13040b79661cbbc8c21febd2f475199e1515c11c