[debugger] Remove SymbolContext from EvalContextImpl.
This data (effectivel the offset of the module in memory) can now be
computed from the symbol data which came from that module. It does not
need to be tracked separately. This part of the change should have no
behavior effect.

Add function addresses to the debug symbol output now that the absolute
address can be computed.

Bug: 41540

Change-Id: I6469b7a4c7f71ce45c13886137d0b26b7f4dfed7