[line input] Remove BeginReadLine
The LineInput class now manages its own state and resets itself when the
user acceps a line input. The embedder no longer needs to manage this
with BeginReadLine() which is now removed.

Convert EOF handling to use a callback to avoid keeping state on it.

Add a simple program which serves as an example and a manual test.

Change-Id: Ib0e82b608be5fd0c9bcf7cfa2d26ffdc1fdc5e59