[debugger][x64] Fix x64 single step issues.
When a thread single stepped, the thread would not set the single step
mode correctly, making it single step twice after going over a
breakpoint. While not noticeable in most scenerios, it would slow down
breakpoint-heavy applications like fidlcat.

Also correctly clears the DR6 register when appropiate. The BS bit is
not cleared by the CPU, but rather it has to be cleared manually. That
means that some HW exceptions could incorrectly be interpreted as single
step.

TEST=Add unit test for bit clearing.
     fx run-test debug_agent_tests

Change-Id: I875ae47796223b37adb3d2c10a884a4f32590121