[fs][journal] add journal parsing test for corruption bug
this tests the same thing that the additional test for the replay tree
is testing, except at a higher level with the journal parsing logic
involved as well, where we aren't manually setting the vmo_offset. this
more clearly shows the exact behavior that caused the original bug,
which is two contiguous write operations that are merged into one and
then inproperly split after the fact. fails without the fix present in
332429.

Bug: 39170
Change-Id: I0a225922346999cfa7cdaf4070afccf7ad207603