[scenic] Always call ObjectLinker's link_failed on Link invalidation.
Prior to this change, when a Link is invalidated, its peer Link will have its
link_failed callback invoked, but the Link being invalidated will not.

While this behavior is advantageous when a Link is being destroyed because the
object that owns it is being destroyed, it is inconsistent with what Links
represent. When a Link is Initialized(), it represents a potential connection
with a peer. Therefore, destroying an Initialized() link means that regardless
of why the Link is being destroyed, that potential connection failed, and
therefore the callback should be invoked.

This patch has two similar-but-subtly-different changes:
1. Invalidating a Link will result in link_failed callbacks to both it and its
   peer.
2. Invalidating a Link that was never connected will still invoke it's
   link_failed callback. This includes the destruction of that Link.

For clarity, link_failed is now called link_invalidated. It is also now passed
a boolean indicating whether or not link invalidation was caused by the
destruction of the link or not to perform the appropriate cleanup (to avoid
cases where the callback accesses data in the object that owns the link).

The peer's link_invalidated callback will always be called before the Link
being closed.

Fixed: 41426

TEST=fx run-test scenic_tests

Change-Id: Ia6e41e04809f4a35c4335482080c9d3634fa9642