[amlogic-decoder][vp9] Implement using parser for input
We need to be able to use the parser to copy protected inputs into the
protected stream buffer. To allow us to continue to use the
multi-instance decoder support, instead of automatically hooking the
parser output registers to the core inputs we need to manually sync
updates between those registers. This also requires that the stream
buffer be big enough to hold an entire frame at once (the same as we do
with the parser disabled).

Testability: VP9 integration tests for low-level parts. The parts in
CodecAdapterVp9 are completely disabled until we get the full
secure_input_output_test working.
Bug: 35200

Change-Id: Ie70c28f9e931204e0ff856f791283c6a50158bf6