[xhci-rewrite] Add synchronous asynchronous executor
This patch adds an executor to the xHCI rewrite, which permits
the use of fit::promise without yielding to the scheduler.
This executor is needed to drive promises to completion
from the interrupt thread. The idea behind this executor
is that tasks running in the xHCI driver will be able to post
deferred work which will be ran later after the interrupt thread
finishes processing events on the event ring, but before
it yields control back to the scheduler (by waiting on the port).

Compared to single_threaded_executor, this implemmentation
is interrupt-safe, meaning that it is safe to invoke from
an interrupt thread context, and will not block unless there
is contention over a mutex).
Under normal circumstances it should not yield the caller’s timeslice.

Change-Id: Ibf543b41ec1997817d817dad96288834412f5fde