[cm] Add support for routing runner capabilities.
Components offering/exposing runners will now be resolved, and their
service be will be opened ready to start a client.

We don't yet fully implement the client side logic. A channel will be
created and passed to a "RemoteRunner" class, but the class does not
yet attempt to use the channel.

Bug: 4761

Change-Id: I14dfb447eff3997cb371496c30c545796c756aa2