[graphics][compute][spinel] Test fixture for Spinel context
This test fixture performs setup/teardown of a Spinel context.

This is the first building block for Spinel unit testing.

The fixture depends on the following test "environments":

  * creation of a VkInstance and supporting Vulkan objects
  * a Spinel target binary
  * creation of a target-specific VkDevice

The first test obsoletes `tests/spinel_vk_context`.

The second test in the suite adds a single path builder test.  This
validates the reuse of the test suite's shared state between tests.

Testability:

  * Target: fx run-test graphics_compute_vk_unittests -- --gtest_filter=spinel_vk.*
  * Host: fx run-host-tests graphics_compute_vk_unittests -- --gtest_filter=spinel_vk.*

Change-Id: Ib1904ac52827f208d258c32fae862a6c11dd9db4