[netstack3] Initial commit of IP sockets
This commit adds initial support for IP sockets. IP sockets define
the IP layer information associated with any IP or transport layer
socket (raw IP sockets, UDP sockets, ICMP sockets, etc).

In addition to storing information that defines the socket, sockets
also cache information that allows the packet send operation to avoid
doing work. Instead, we push this work to the update path so that
updates (like modifying the forwarding table) must update all sockets
in order to keep their cached information valid. This is the price
for making the packet send operation as fast as possible.

As usual, we introduce traits for sockets and for an execution
context which uses a particular type of socket. This allows us to
use a test mock (implemented in the `testutil` module) when testing
a transport-layer protocol that doesn't want a real IP layer.

Test: test_new, test_send

Change-Id: I8a28e022ae7149c111ec326e19049c71976be6fe