[debugger] Get the module from a unit.
Store the ModuleSymbols object associated with a compilation unit. This
allows querying for the module for any symbol since every symbol can get
back to its compilation unit.

Adds a way to look up a loaded instance for a given ModuleSymbols
pointer in a process. This allows a Symbol object to compute its
SymbolContext if given a process.

This is currently unused but will be hooked up to some places that need
this information in a followup.

Change-Id: Id1b213c9eb1bc344ab6f29583777d60479da36fc