[kernel] remove port-based exception syscalls
Removes the deprecated port-based exception syscalls. This just deletes
the API to start with so we can detect and rollback if there are still
clients using it that I've missed.

If this goes through without triggering any build failures, I'll post
a follow-up CL to remove all the corresponding implementation.

Bug: 33817
Change-Id: Ifae524dd8110a9872696d3ad4a87c9670ce7bfc3