[perfcompare] Print results as decimal fractions, not just integers
Before this change, perfcompare only printed the integer part of
result value confidence intervals.  That worked OK for nanosecond
times, e.g. "7656 +/- 1275 ns", but not for millisecond times, which
would often print "0 +/- 0 ms".

Change that to print decimal fractions.  Print a number of digits
that's appropriate to the width of the confidence interval,
e.g. "0.0657 +/- 0.0040 ms".

Note that some of the golden file output changes because previously we
used "%d" to format the numbers, and that truncates (e.g. 1.6 becomes
1) rather than rounding to the nearest integer.

Bug: 41460
Test: "python perfcompare_test.py" (run by bots on CQ) + perfcompare trybot
Change-Id: Ie0e238a21f7909c7889fbe450f2fe10ad9c337b4