[fx reboot] best-effort exit with success
This makes `fx reboot` more useful in command chains, so you can do things
like:
  fx reboot && fx wait && fx pave -1

The implementation is a little weird, because it has to overcome some fiddly
race conditions. It was tested by replacing the reboot with echo (to test the
normal "failed to reboot" condition), and by inserting a "sleep 20 &&
fx-command-run shell echo" between the reboot and the loop for the other
possible race condition.

Change-Id: If4beda8476bbe04a4e2a130b9175ea24d474b678