[session] Don't launch session by default
This makes the default behavior of session manager be to wait for a
session to be launched from the command line.

A default session can be added back when launching a new one actually
terminates the old session, but for now it's confusing to have say a
graphical session start when you launch a session from the command line.

Change-Id: I005b0d5aa08c80c6ef1c822db9a01e0c6309c9a5