[pkg-resolver] Handle http 429 responses
When fetching blobs from the network, a web server may respond with "429
too many requests".  In that case, the resolver was retrying the blob
download immediately once (as it does for all network related errors
when fetching blobs), but the intent of the 429 status code is to rate
limit the client to slow down their retry attempts.

This change handles the 429 response specially by exponentially backing
off its retry rate for that blob and increases the number of retries for
this error to as many as can fit in 15 seconds. Additionally, the
package resolver now tracks network flakes and 429 responses per-mirror
in its inspect metrics.

Fixed: 37164
Change-Id: Id731cd9ff69254339739569408a2a59eb5e11988