[component] Zero-initialize response buffer
UBSan was reporting a load of an invalid boolean value (170). This
indicates an uninitialized boolean value being used. The source was from
an uninitialized char buffer in DdkRxrpc() which would copy it's value
into `ClockProxyResponse::is_enabled`.

This could also potentially affect fields in other responses, so we can
just zero-initialize the response buffer to avoid this.

The impact of zero-initing this buffer is minimal. Because we are
already using -ftrivial-auto-var-init=pattern, the only difference in
assembly is:

  < 	xorl	%esi, %esi
  ---
  > 	movl	$170, %esi

Fixed: 41875
Change-Id: Ib7e1cfb6780b7d046420e8e9fc7cbc37b099274e