[blobfs] Do not leave VMO attached
This is an aid for hunting fxb/36257 and should be reverted once that
is resolved.

Currently the mapping_ VMO is only directly read from during InitVmos,
and this is the only time it needs to be attached. In order to limit
the chance of the VMO being accidentally used by either the block
server, or elsewhere in blobfs, this change does not store the vmoid
and immediately detaches the vmo from the block server.

Bug: 36257
Change-Id: I791731cfa7a66a0947f7fb573bd8533ef84d18e9