[blobfs] Perform writes from clone
This is an aid for hunting fxb/36257 and should be reverted once that
is resolved.

Performs writeback using a clone of the mapping_ VMO to avoid the need
to send expose the mapping_ handle. Aside from attaching the VMO to the
block server, this represents the only location where the mapping_
handle 'escapes' and becomes used by other code. The journaling path
should not be happening in parallel to reads, nor should it be pinning,
but this makes it easier to reason about who could have access to the
VMO.

Bug: 36257

Change-Id: Idab8a2a5e610a51e94dddd02a86f079b6610e9cb