[debugger] Separate out dwarf expression eval.
The EvalContext had a helper class internal to it that helps manage
evaluating a DWARF expression asynchronously. There are some other cases
where we will need to do this same operation, so this moves out the
helper and the conversion of the result to an ExprValue into a separate
class.

Additionally, EvalContextImpl now computes the SymbolContext for the
expression it passes to the evaluator from the variable that had the
expression. This is important for cross-module global variables.

In a followup, the symbol_context member of EvalContextImpl will be
removed.

Change-Id: I3af9d21dc5dd77bac73bb2bda634d7dd172c4985