[ftl] Add read support for FTL disk format 2.0
The FTL library is now able to read and write NDM control headers with
explicit version information (aka, "2.0"), but will still only write
old headers by default (aka, "1.1").

The base NDM driver is now able to save configuration data to the
partition table, so that we can compare what a new software version wants
to do versus what the code that created the volume wanted to do.

This CL adds a new FTL test executble (ftl-lib-test), for unit tests
of the FTL library.

The version of the FTL driver is bumped to 1.2.0 (informational only).

Bug: 40208

Change-Id: I1662afa2ab8ce001b9d46669b9cb42fb5ba33f0c