[camera] speed up stream_provider_test
This change updates the stream_provider_test to hash a fixed subsampling
of points within an image rather than the entire image. This speeds up
the test significantly, reducing the time frames are held from 100ms to
8ms, measured using added trace commands. This change also re-enables
the MANAGER path, updating it to be an injected vs. environment
component, and allowing it to be skipped if in use.

Change-Id: I17a9afb7db0b16c8e482fba89f10630ccbc5b0f6