[SetUi] Load default config in Device Storage
Add functionality in Device Storage so that it can
load default settings from a configuration file.

Replaced the ConfigData trait with a DefaultSetting
struct, which would load and cache the config data
if a path to a file is provided.

Modified Device Storage so that it'll store a
DefaultSetting struct instead of a default value.

Bug: fxb/41938
Tests: Modified the tests in loader.rs (now
   default_setting.rs) so that it tests for the
   Default Settings struct
 Run with fx run-test setui_service_tests -- --test

Change-Id: I025a3f26ebb58c2431e807fd454869ed30e0242d