[fidl][overnet] Move host-side handle code into fidl crate
When I started authoring this code, I expected that there'd be a strong
Overnet dependency from the FIDL support code, and to mitigate this we
used FFI to break the dependency cycle that would have resulted.

Upon inspecting the code today, I realized that no such Overnet
dependency exists from this code, and we can instead fold all of the
implementation into the FIDL library.

This has the advantage that Overnet libraries need not be mentioned by
host based applications that happen to want to serialize FIDL messages,
in addition to eliminating the bulk of the unsafe code in the libraries
implementation.

Also corrects a bug (and adds a test to verify the fix) whereby a socket
handle would report its type as channel.

Change-Id: I4a771ab297d5543cb58fc1de7ea57463bcada89e