[debugger] Fix use-after-free when walking up the symbols chain.
Symbols' parent are "uncached" and the getter returns a RefPtr to manage
the memory of the returned object. This was a recent change and not all
users of the parent() getter had been updated.

Audits callers of parent() and adds RefPtr to hold the lifetime of
the resulting objects. Extend to GetContainingFunction() and
GetInlineChain() which both need to return references.

Convert Function's containing_block() to use an UncachedLazySymbol since
it's an upward reference in the symbol tree. This is equivalent to the
parent() usage but this case was missed when parent() was converted.

TEST=These callers have tests but the tests didn't catch the
use-after-free.

Change-Id: I32236a797dc7cdc0072f4c436dc98d310b815c5a