[zircon][debugger][x64] Correct DR6 handling.
This CL addresses the issue with DR6. The Single step bit is not cleared
automatically by the CPU, but rather has to be done manually. Before
this responsability was left to the user space program (debug_agent).
But in reality DR6 was not accessible.

This change moves that responsability to Zircon, making the "ephemeral"
bits on DR6 be valid only while the exception handle is valid. When it
returns, those bits are cleared and user space programs can read DR6
correctly.

This was interfering with the correct usage of HW breakpoints. This was
because the debugger was "stepping on itself", clearing hw breakpoint
control definitions when attempting to clear those bits or write new hw
breakpoints definitions.

This CL also adds kernel debug functions for printing DR6 and DR7.
This CL also fixes a bug in the integration tests that would make the
tests hang under certain conditions (the { } around the log statements).

TESTS=All current tests pass.
      Manually tested stepping, sw breakpoints and 3 hw breakpoints.

Change-Id: I3836c26921384510fa313f2892b587179a1719ce