[netsvc] don't initialize debuglog in limited mode
This change prevents a pathological case introduced with limited mode in
e6fbe21. That change disabled the emission of debug logs for user
devices; unfortunately, it did not disable the initialization of the
debuglog object in netsvc. This caused the debuglog deadline to be
immediately lapsed, and the disablement of periodic log emission
prevented the deadline from ever being advanced.

Conditioning debuglog initialization on full mode operation fixes the
problem.

I suspect this may be related to issue 41248.

Bug: 41248
Change-Id: I256e8bad1e4319ccebc843d37e26a16825b6c381