[fidl][HLCPP] Optimize validation for v1 messages
When transforming message bytes from v1 to the old wire
format for validation, HLCPP will now either stack or heap
allocate memory to contain the transformed message depending
on its size, similar to LLCPP. Also, if the message's v1 and
old wire formats are the same (ie it does not contain any
unions) the transformation is skipped altogether.

In order to achieve this, the max_out_of_line and
contains_union fields are added to the struct coding tables.

This CL also contains some additional changes to make CQ
green when v1 is enabled (e.g. adding a wait to bus_unittest.cc
to reduce flakiness).

Change-Id: Id81cf0d1d27f3b6d3863d070f6927776f3e260c7