[fdf] Add InitTask.
This is for ensuring a device will not be unbound until
the init hook completes.

Devhost side changes left to future CL.

In future the devhost will pass a flag to AddDevice indicating
whether the device has an init hook. If so, the coordinator
will create an init task and add the device as invisible.

Once the init task completes, the coordinator will set the
device as visible.

Change-Id: I37e312116f6197b460d72427a3c5cbc849523902