[audio_core] Refactor for usage-based routing
This CL refactors RouteGraph to prepare for usage-based routing.

Currently we depend on the `deque` alone for choosing targets, because
we always route each stream to the last-plugged. Implementing usage
based routing requires we have additional accounting to know targets for
each usage.

This CL introduces the additional accounting to implement the same
behaviour so that the CL introducing usage based routing has a coherent
diff.

No behaviour is changed.

Bug: 34922
Test: fx run-test audio_core_unittests
Change-Id: I47a2deea794b86d527398139e52efdcec5583267