[display][intel] Make type argument optional when setting sysmem constraints.
Vulkan drivers provide this constraint and WSI code typically
don't know what tiling format is preferred by the Vulkan
implementation.

This change makes the image type an optional configuration. Setting
it to 0, tells the display driver that the buffer collection
should not be constrained to a specific tiling format. This makes
it possible for a different participant (ie. the vulkan driver) to
provide this constraint.

Test: fx run display-test
Change-Id: Ic3fdcfd1456dd0daad931d75837273987bf229f0