[audio_core] Stop Capturer on FIDL channel drop
This CL ensures the AudioCapturerImpl drops when its FIDL channel
closes.

First, it synchronizes with the mix thread to ensure the mix thread's
`fbl::RefPtr` is dropped, and `fbl_recycle` is triggered.

Then, whether AudioCapturerImpl's `fbl_recycle` runs
depends on the type of the last reference dropped. In tests, the last
reference was of type `fbl::RefPtr<AudioCapturerImpl>`, so our
`fbl_recycle` would run.

In production, the last reference was of type
`fbl::RefPtr<AudioObject>`, so our `fbl_recycle` would not run, leaving
the capturer up to spew logs and hold resources indefinitely.

Per 39624, we should move to `std::shared_ptr` soon so that we can
use bound deleters instead.

Testing: Logging is turned on in the tests so that the DCHECKs in the
destructor will fire if Shutdown() is not run first. Another DCHECK is
added.

Fixed: 42077
Bug: 39624
Test: fx run-test audio_core_unittests
Change-Id: Ieec90318e59af1b0807c727bc575a5d82fc539ab