[identity] Traitify AccountHandlerConnection
Make AccountHandlerConnection a trait, and let the production
implementation of it be called AccountHandlerConnectionImpl.

AccountHandlerConnection is no longer responsible for initialization
(create and load account), but only to establish the actual connection
which requires an account id, an AccountHandlerContext, and the
lifetime. Initialization is moved out to their respective locations in
account_map and account_manager.

Finally, account_map is responsible for minting new connections,
which allows it to perform collision detection for new account ids.

Bug: 37433

Change-Id: I32e66e0b6519f0bdf44512da540dc2a418f26f3f