[identity] Remove duplicate impl on FakeAuthProviderSupplier
This change removes the duplicate implementation of
the AuthProviderSupplier trait on references to
FakeAuthProviderSupplier. The duplicate impl was
added to support tests where it was necessary to
hold a reference to the fake in one task, while
moving ownership of the fake to another task.
This is resolved by removing the reference in the
first task.

This change also removes a check that verifies
that all registered clients are used.  The failure
mode in this case, both before and after this change,
is that the test stalls or times out, and therefore
the check is not invoked in this case.

Testing: Existing unit tests

Change-Id: I6809a41523754bcb4f5aec0e4ff2009447dcf834