[ledger] Get diff in PageStorage
This enables getting diffs in PageStorage: when downloading a piece that
is part of the tree of a commit, the piece itself is requested from P2P,
while a diff to this commit is requested from the cloud.
In compatibility mode, we assume that the parent commit is not
necessarily locally present: we may have to download from the cloud. It is
downloaded as pieces to avoid requesting a diff again. If a commit does
not have an associated diff, or if the cloud does not support diffs,
stroage will receive an empty diff from the same commit, which will be
handled by the same fallback strategy.

The P2P and cloud download behaviors diverge significantly: some sync
related tests in page_storage_unittest.cc are parametrized by the
available synchronization mechanisms, and by whether the compatibility
behavior is enabled.
Integration tests also run with compatibility disabled, and with
compatibility enabled and randomly missing diffs (for most tests we run
both. PageSnapshot tests are particularly slow and should not be
affected by diffs so we choose one of these modes of operation at
random). To avoid timeout due to the duplicated tests, integration tests
are sharded between tests without compatibility enabled and tests with
compatibility enabled.

MULTIPLY: `[
  {
    "target": "ledger_integration_tests",
    "os": "fuchsia",
    "total_runs": 4
  },
  {
    "target": "ledger_integration_tests_diff_compatibility",
    "os": "fuchsia",
    "total_runs": 4
  }
]`

Change-Id: I53a1dce9ba805dfc6ebde524e4f6d8f6d914765d