[i2c-hid] Fix UBSan error
UBSan is complaning about accessing a member of an object from an
address that has inssuficient space. The issue is that we're casting a
uint8_t[4] to a I2cHidDesc which is larger, so UBSan complains that a
potential member access into this casted buffer could access some
uninitialized data, which is undefined behavior. This particular member
access should be ok though since the field we are accessing (wHIDDescLength)
is the first member of the struct.

We can cast to a uint16_t* then dereference to get this data.

Fixed: 41880
Change-Id: I09c07804dca7bb676f344bc83fc616718279590f