[symbolize] Restart process to reduce mem usage
This change causes symbolize to restart llvm-symbolizer regularly in order
to reduce how much memory is consumed by it. llvm-symbolizer holds on to
a lot of data as the process runs which makes it non-ideal to use for
cases like this where it sees many new binaries.

Test: Logged into a bot and ran many instances of symbolize while monitoring
top to check on memory usage. Memory usage was quite low with these changes.

Bug: 41921
Change-Id: Iedd45a8f9e5f340b37ced1d5db122b1ea5aa35ee