[sl4f] Add RequestId/MethodId types
As an extension to the json-rpc protocol, sl4f supports storing client
metadata when a request ID contains a '.' character, interpreting the
first part of the request ID as the client identifier. Unfortunately,
this feature also strips the first part of all request IDs, so clients
not aware of this extension will not observe responses with the expected
ID field.

This change modifies sl4f's internal handling of request IDs and method
IDs to use dedicated types with their own parse function, extending the
handling of request IDs to allow arbitrary json types (golang, for
example, expects this field to be an int), and to not strip out the id
if it does not contain a '.'.

Test: act.py -c <config> -tc GattServerSetupTest
Change-Id: I96da9bd9185f1b0f7f721d98a345f9de044dfdbf