[SetUi] Audio default setting config
Load default audio settings via a config file.

Introduced audio_default_settings.rs which contains
the hardcoded default values and a global Default
Settings object for AudioInfo. The object is global
to prevent the config file from being read multiple
times.

The Device Storage is modified so that it now
takes in a default setting value instead of a Default
Setting object.

AudioController is modified so that it takes in the
default audio settings from the audio_default_settings

Tests: Added a test in audio_default_settings.rs that loads
 a config file for AudioInfo.
 Modified the existing audio tests so that they would
 load default values from the config instead of using
 the hardcoded values.
 Run tests with:
   fx run-test setui_service_tests -- --test
Bug: fxb/38447
Change-Id: I9c88f777ed23299e4fa685b7a98f746d17d01498