[debugger] Allow whitespace in certain commands.
Certain commands take expressions or other input which can contain
whitespace. An example is the "print" command. If you type

  print x + 1

it means to print the expression "x + 1". The previous command-line
parser split all input on whitespace so this have three arguments, while

  print x+1

would have one argument. This required putting back the command-line
with whitespace, and will break when language string literals are
present.

This change marks these commands as taking one parameter rather than
parsing the arguments as whitespace-separated. The parser is updated to
produce a single argument string in this case.

Change-Id: I333494c4dee26eaab88e480396cd2b64f8c15737