[build] Clarification of names of rustc_binary outputs
The rustc_binary() template creates a number of outputs that have names
based on transformations of the name passed to the template.  Either by
adding underscored suffixes, or by converting dashes to underscores.

This change adds language to the documentation for the template spelling
out what those transformed outputs are, and includes an additional
example that shows how they look when used.

Test: fx build
Change-Id: I81dba9a5e4eb021e83e6699da3c4ca508e610e96