[cm] Add tests for using invalid cap from component manager.
Previous versions of this code attempted to unwrap the current moniker,
but because it was at the component manager (which has moniker == None),
a panic was triggered.

Change I55a27889945c5d51d2dbf10aed85dcbc9bfbfd59 fixed the bug (in
parallel with this CL :-). This CL adds tests exercising this path.

Bug: 42160

Change-Id: I60abd46899e8fc1c885445c1445fa434db066480