[wlanif] SoftAP crashes when operating in protected mode.
Bug: 42068

Use the vector constructor that copies from an array using its length.
Added a convert routine to do the conversion to fidl and added
a convert test.

Test: Manual testing of starting an AP in protected mode and
associating to it from an external client.

Also verified the newly added sub test, ToFidlAssocInd passes
with this change (and crashes without it).

Change-Id: Ia12b73b1c907b5c9d968056c0c7575a30901338f