[cm] Simplify building of directories in OutDir.
`OutDir` is used to construct the outgoing directory of test components.
It's current implementation is hampered by `DirectoryEntry` being
inherently single threaded, and not being clonable.

We introduce a new class `SyncDirectoryEntry` that is a Send + Sync
+ Clone wrapped around a `DirectoryEntry`. Using this, we can simplify
the implementation of `OutDir` to no longer need to construct new
directories on the fly based on Cloneable data, but just take input
`DirectoryProxy` and `DirectoryEntry` objects directly.

These changes are to faciliate testing of CFv2 runners in upcoming
CLs.

Bug: 4761
Change-Id: Ia55dfea6c09eb6b4962f97cafc8dc04d9f706ca8