[cm] Better encapsulate MockRunner's internal state.
This change attempts to improve encapsulation of internal data of
MockRunner. Currently, users directly access fields of the MockRunner
struct, such as cloning references to the "namespaces" map or "urls_run"
set.

We update MockRunner to no longer mark internal state as public, and
modify callers to use MockRunner's interface.

In preparation for further cleanups.

Change-Id: I7b93767e5d836b1f42263120def801da9a6e4e2b