[fbl] Speed compilation of tests by avoiding use of compile-time sizes.
The vector tests currently run on a variety of vector sizes, each of
which gets its own copy of the function. This leads to a large amount
of code being generated, leading to compile times of ~3 minutes.

This CL updates the code to use dynamic (i.e., runtime) sizes, reducing
the code size (and hence compilation times) by a factor of 6.

Change-Id: Ic1aae4d7da7bf37536d5eb8aaf03c36793ebf880