[blobfs][pager] Mount option to enable paging
Paging is off by default, but can be enabled by setting |pager| to true
in the MountOptions struct passed in to |Blobfs::Create()|. Added
commandline option '--pager' to do so.

Once the pager is on by default, we can switch this mechanism to turn
off the pager if needed.

Test: Added commandline options to enable the pager in blobfs
integration tests: blobfs-test, blobfs-integration-test,
blobfs-large-integration-test, blobfs-bench-test.

Change-Id: Id99e8046fd4a4d5e358103b8d5b12e6f84ec55dc