[blobfs] Add info_mapping_ null check in Blobfs::Reset
When Blobfs::Create call fails, it is possible for
info_mapping_ to be uninitialized in the half-created blobfs
Object. In the destructor, the Reset method is called which
calls info_mapping_ to perform Journal clean bit writes. Added a null
check in Reset to only perform clean bit write if info_mapping_
is initialized.

Bug: 39510, 42174

Test: Fixes AddressSanitizer errors in TestCorruptMount in
utest/fvm/fvm.cc. Passes "runtests -t blobfs-test"

Change-Id: I14a8c91db208123d110a7493b05c04e842bd938f