[run_test] Swap loop and ComponentContext creation
The normal idiom is to create the async loop first and then create the
component context. It's unlikely that the previous code, which
initialized these objects in the opposite order, had an actual bug, but
it's an error-prone pattern. This CL switches the code to the normal
pattern so that future people using this code as a model end up with the
normal idiom.

Test: No behavior change
Change-Id: Iea4f191b854318c3d4aec40655a043fdd7c9046c