[debugger] Improve "set" command.
The documentation said that lists were colon-separated but this was a
lie. Standardizes on whitespace-separated lists with quoting if
necessary.

Allows whitespace or no whitespace on either side of the operator, as in
"set foo = bar" and "set foo=bar". Previously whitespace was required
which was surprising.

Improves messaging for modifications to just include the full new value
rather than the values added or changed. This helps the user understand
the current state.

Add the setting name to the output of the set command.

Update the help to give real setting examples now that we have more.

Change-Id: I605b87bd10c5260beaf3724b2edb74901e61f080