[debugger] Convert some observers to global.
Previously target/process/thread observer lists were on each
target/process/thread. But currently almost all users in our system want
to observe everything. This required a lot of extra tracking for each
object in order to be able to install and remove the observer
registration for each one.

Converts the observers for these objects to be global on the session.
This simplifies most code that uses observers.

Moves the create/destroy notifications for each object type to the
observer for that object type, rather than its parent. This helps
clarify things.

If we need per-object observers for specific cases (probably for
performance reasons) in the future we can add those observers in
parallel.

There should be no behavior change. Most changes are trivial updates,
the only nontrivial changes are in breakpoint_impl.cc which did a lot of
tracking for observers, and to a lesser degree console_context.cc

Change-Id: Iec2f658dfea5c289c1fda8440a3cb04cc8c6928d