[identity][sktool] Introduct abstraction layer for device connection.
This adds a simple "connection" trait with a FIDL based implementation
that is used in real life and tested with a proxy, and a Fake
implementation that can be used for testing across the rest of the tool.

This wasn't essential now but the proxy-based testing approach would
have got very cumbersome when we need to start testing and generating
package chains.

This also allowed one or two small minor cleanups.

Bug: 40981
Change-Id: Ide5a75cd10e731341849e8d0f622c1734b17417b