[build] Clarification of names of rustc_library outputs
The rustc_library() template creates a test binary that has a name
based on transformations of the name passed to the template, by
adding underscored suffixes, and by converting dashes to underscores.

This change adds language to the documentation for the template spelling
out what the transformed output name is, and includes an additional
example that shows how it looks when used.

Test: fx build
Change-Id: Ie558db74ddf43dd9ab8259ba6b1781efd6af48b5