[graphics][compute][spinel] Enable render clip
Bug:25525

The clip used by the render shader was not in tile coordinates and was
too large in the `spinel_vk_svg` test.

Currently, similar logic for converting pixel clips to tile clips
appears in both the composition and render host-side implementations.
These will likely diverge in the near future.

The clipping logic was unchanged in the shaders but the PLACE_TT*K
shaders will receive updated logic in a future CL.

The unit test demonstrates that both the composition and render clips
have the same effect.

The integration tests were updated to default to a clip that's equal to
the maximum surface size.

Dead code in `vk_layouts.h` was also removed when updating the clip
name.

Tricium required formatting s/g/l/c/spinel/spinel-rs-sys/src/lib.rs.

Testability:

  * Target:
    fx run-test graphics_compute_vk_unittests -- --gtest_filter=spinel_vk_svg_*.svg_tests/*_clip

  * Host:
    fx run-host-tests graphics_compute_vk_unittests -- --gtest_filter=spinel_vk_svg_*.svg_tests/*_clip

Change-Id: Ic53bfd922a4a6881503cb8ce1eba93a8a9a49075